CHET-623: Refactor the migration workflow to separate pages
