+++ jira-e2e-tests/cypress/support/pages/LandingPage.ts
@@ -1,0 +1,7 @@
+export const startMigration = (ctx: AppContext) => {
+    cy.visit(ctx.pluginHomePage);
+    cy.location().should((loc: Location) => {
+        expect(loc.pathname).to.eq(ctx.context + '/plugins/servlet/dc-migration-assistant/home');
+    });
+    cy.get('[data-test=start-migration]').should('exist').click();
+};
